refactor: folder capitalization and lint error
