+++ libs/journeys/ui/src/components/Video/VideoControls/utils/playbackReducer/index.ts
@@ -1,0 +1,5 @@
+export {
+  playbackReducer,
+  type PlaybackState,
+  PlaybackEvent
+} from './playbackReducer'
